porca madonna

+++ Firmware Sistema/Sistema Pylon Rev2 18-07/Palo 2e3/RXROSSO_Palo2-3_Luci/RXROSSO_Palo2-3_Luci.ino
@@ -41,7 +41,7 @@ String PressCut3 = "3311";
 String State=Startup;
 unsigned long Delaysend=200;
 unsigned long DelayState=350;
-int Chan=9;
+byte Chan=9;
 //Fine Configurazioni Ricevente
 
 void setup() {
